+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 * create `ItemLayoutWrapper` for easier wrapping of an `Item` inside a different layout
 * add `HasWrappedItem` interface (2 implementations - `ItemWrapper` and `ItemLayoutWrapper`)
 * utility to automatically process nested RecyclerView state - `NestedRecyclerState`
+* add `Adapt#getItem(position)` method
+* add `AdaptView#view()` method
+* add `StickyItemDecoration` for sticky headers/sections